pyproject.toml
[project]
name = "programador-turnos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.1",
    "ortools>=9.7.2996",
    "pyyaml>=6.0"
]


